BlackJack Game
A Blackjack game with randomly generated playing cards numbered from 1 to 13 with Ace, Jack, King and Queen numbered as 1, 10, 11 and 12 .
Starting the game 2 cards are dealt. If the sum is 21 , player got blackjack.
If the sum more than 21 , player is out of the game.
If the sum is less than 21, player can either restart or draw a new card until getting blackjack or out of the game.